refactor(cli): isolate workflow progress projection
